Settings, buttons, and listeners
Added logic to the FragmentSong buttons except for loop and repeat.
Seekbar is fully functional.
Moved code that only needs to be ran once in ActivityMain.
Added a fab to FragmentPlaylist, but need to make it show().
FragmentSongs closes keyboard.
Added settings to change the ProbFun parameters.
Added OnPreparedListeners and OnErrorListeners to the MediaPlayersWURI.
Cleaned up the GUI.