Bug fixes
Now update seekbar when song is in progress and app is reopened.
Now show song pane when song is in progress and not playing.
Added Bitmaps to 3 image buttons in FragmentSong. They removed artifacts.
Seeking when paused now plays when seeking is done.
Now clear the queue when the user picks a new song.