AddToPlaylist functionality
Each item_song now has an add to playlist button that opens an AddToPlaylistDialog.
UndoListeners for swiping playlists and songs are now fully functional.
If no song is selected on startup and play is clicked, then a song is played from the masterPlaylist and the masterPlaylist is continued to be picked from.