Code clean up and some updates to the ProbFun
Reorganized code by fragments and playlist.
Chaged good in prob fun to scale the percent down when needed to avoid a ceiling on the probability.
Fixed ProbFun's clone method.
Fixed lower probs to lower probs by a lower percent to avoid massive difference in the low prob cutoff and the probs lowered.