Refactoring settings
Put settings in a SettingRepo.
Took them out of MediaData.
Added the SettingsRepo directly to the ProbFun to avoid passing variables.
SeekBar funtionality is broken.
Added a ViewModelSettings class to hold the settings, and started to make FragmentSettings dumber.
Fix bug where the AudioUri was not updated after a new song was supposed to play.