Added a NavUtil function that takes a Fragment id and prevents navigating to the same destination. This allows removal of the old navigation method in ActivityMain.
Renamed the MediaController to MediaModel due to it not being a controller in a programmatic sense.

Added a LiveData to ViewModelActivityMain for the song pane visibility.

Moved FragmentPaneSong's onDestinationChangedListener to ActivityMain.

Tried to remove the Context from the MediaModel by passing it through methods. Unfortunately, it is still in the onCompletionListener.