ActivityMain.kt cleaned up.
Moved Toolbar title TextView formatting to XML from ActivityMain.kt.

Now use a PorterDuffColorFilter for all APIs when setting the overflow icon color in the Toolbar.

Removed unnecessary actions from the BroadcastReceiver in ActivityMain.kt.

Removed the broadcast from ActivityMain.kt that was used to let Fragments know that the options menu was created, as they now use the onPrepareOptionsMenu callback.

ActivityMain.kt now adds the non-null song or playlist to the queue no matter what when the user clicks addToQueue.

User can now add a playlist to a playlist via the Toolbar.