Added a UI for the queue with sorting and removing functionality.
Added an icon to open FragmentQueue.kt when there is a queue.

Added FragmentQueue.kt which uses RecyclerViewAdapterSongs.kt to allow the user to interact with the queue.

Fixed bug where app crashed if user did not pick a folder when searching for one.

SongQueue.kt now holds Songs rather than the ids, allows clients to get the queue as a Set, to move songs, to remove songs, to insert songs, and to start playback at a particular position in the song queue.