Bug fixes related to last commit.
The action_queue icon now shows on the Toolbar after the MediaSession is started.
FragmentQueue.kt updates its RecyclerView when the queue changes.
FragmentQueue.kt now stops the current song if it is removed from the queue and starts playback of the next song.
SongQueue.kt now exposes its queue through LiveData.
Fixed bug when the current song was moved, but the queue iterator was not updated to its new position.
Fixed bug where if the last song in the queue was playing and removed from the queue, it threw and exception. Now it continues playback.