Did some refactoring; untangled a few dependencies.
MediaDatasource now loads the database on instantiation.

Changed the loading screen to a three step process

Lots of changes were made to ProbabilityFunction.

Fixed the bug where updating the Notification resulted in a TransactionTooLargeException due to a memory leak in the RemoteViews. This resulted in the RemoteViewCreator. Issue is described in detail here: https://stackoverflow.com/questions/22789588/how-to-update-notification-with-remoteviews/53222877#53222877